+++ src/renderer/index.tsx
@@ -2,6 +2,7 @@ import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import { AlertsBar } from "./ui/dialogs/alerts/AlertsBar";
 import { Dialogs } from "./ui/dialogs/Dialogs";
+import { Progress } from "./ui/dialogs/progress/Progress";
 import { Poppers } from "./ui/dialogs/poppers/Poppers";
 import { MainPage } from "./ui/app/MainPage";
 import "./editors/register-editors";
@@ -13,6 +14,7 @@ export default function AppContent() {
             <GlobalStyles />
             <MainPage />
             <Dialogs />
+            <Progress />
             <AlertsBar />
             <Poppers />
         </DndProvider>